third_party/blink/web_tests/external/wpt/css/css-values/tree-counting/calc-sibling-function-expected.txt

This is a testharness.js-based test.
[FAIL] sibling-index() and sibling-count() on nested pseudo-element
  assert_equals: expected "4" but got ""
Harness: the test ran to completion.
